admin_panel/backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Storage & Files
boto3==1.34.34
python-magic==0.4.27

# AI & Embeddings
openai==1.10.0
tiktoken==0.5.2
pgvector==0.2.4

# Utilities
httpx==0.26.0
aiofiles==23.2.1
python-slugify==8.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
